+++ package.json
@@ -67,13 +67,12 @@
     "events": "^1.0.0",
     "hdkey": "^2.0.1",
     "https-browserify": "0.0.1",
-    "jest": "^28.1.3",
     "litecoin-bip84": "0.0.5",
     "litecore-lib": "^0.13.22",
     "lucid-cardano": "^0.7.9",
     "mnemonic-to-key-pair": "^1.0.0",
     "mnemonic-to-private-key": "^0.4.1",
-    "mocha": "^10.0.0",
+    "mocha": "^10.2.0",
     "node-fetch": "^3.3.0",
     "punycode": "^1.2.4",
     "querystring-es3": "^0.2.1",
@@ -120,7 +119,9 @@
     "@babel/core": "^7.18.13",
     "@babel/preset-env": "^7.18.10",
     "@babel/preset-typescript": "^7.18.6",
+    "@types/jest": "^29.5.6",
     "esbuild": "^0.15.5",
+    "jest": "^28.1.3",
     "simple-bitcoin-wallet": "0.0.9",
     "typescript": "^4.7.4",
     "webpack": "^5.74.0"
